+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- **-immutablePrefix** can be set many times and used as a slice.
+
 ## [0.5.0] - 2020-12-28
 ### Changed
 - Use a single out connection and pass packs as pointers to array.